+++ backend/src/main/java/cloud/weareithero/api/asn/dao/AsnDao.java
@@ -1,25 +1,33 @@
-package cloud.weareithero.api.inbound.dao;
+package cloud.weareithero.api.asn.dao;
 
 import java.util.List;
 
-import cloud.weareithero.api.inbound.dto.AsnDTO;
-import cloud.weareithero.api.inbound.dto.AsnMaterialDTO;
-import cloud.weareithero.api.inbound.dto.AsnOrderMaterialDTO;
-import cloud.weareithero.api.inbound.dto.AsnRequestDTO;
-import cloud.weareithero.api.inbound.dto.AsnSummaryDTO;
-import cloud.weareithero.api.inbound.dto.AsnSupplierDTO;
-import cloud.weareithero.api.inbound.dto.AsnWarehouseDTO;
+import cloud.weareithero.api.asn.dto.AsnDTO;
+import cloud.weareithero.api.asn.dto.AsnMaterialDTO;
+import cloud.weareithero.api.asn.dto.AsnOrderMaterialDTO;
+import cloud.weareithero.api.asn.dto.AsnRequestDTO;
+import cloud.weareithero.api.asn.dto.AsnSummaryDTO;
+import cloud.weareithero.api.asn.dto.AsnSupplierDTO;
+import cloud.weareithero.api.asn.dto.AsnWarehouseDTO;
 
 public interface AsnDao {
 
   public AsnSummaryDTO findSummary(AsnRequestDTO asnRequestDTO);
+
   public List<AsnDTO> findAll(AsnRequestDTO asnRequestDTO);
+
   public AsnDTO findByAsnId(int asnId);
+
   public List<AsnOrderMaterialDTO> findOne(int asnId);
+
   public AsnDTO add(AsnDTO asnDTO);
+
   public int addOrderMaterial(AsnOrderMaterialDTO orderMaterialDTO);
+
   public List<AsnSupplierDTO> findBySupplier();
+
   public List<AsnWarehouseDTO> findByWarehouse();
+
   public List<AsnMaterialDTO> findByMaterial();
-  
+
 }
