+++ backend/src/main/java/cloud/weareithero/api/asn/dto/AsnWarehouseDTO.java
@@ -1,4 +1,4 @@
-package cloud.weareithero.api.inbound.dto;
+package cloud.weareithero.api.asn.dto;
 
 import lombok.AllArgsConstructor;
 import lombok.Data;
@@ -7,10 +7,11 @@ import lombok.NoArgsConstructor;
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
-public class AsnMaterialDTO {
+public class AsnWarehouseDTO {
 
   private int id;
   private String name;
-  private String alloyType;
-  
+  private int scale;
+  private String type;
+
 }
