+++ backend/src/main/java/cloud/weareithero/api/inbound/dto/InboundRequestDTO.java
@@ -1,0 +1,5 @@
+package cloud.weareithero.api.inbound.dto;
+
+public class InboundRequestDTO {
+    
+}
