+++ backend/src/main/java/cloud/weareithero/api/inbound/service/InboundService.java
@@ -1,0 +1,11 @@
+package cloud.weareithero.api.inbound.service;
+
+import cloud.weareithero.dto.ResponseDTO;
+
+public interface InboundService {
+    
+    public ResponseDTO findAll();
+    
+    public ResponseDTO findOne(int asnId);
+
+}
